backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
openai
python-dotenv